updates from shawn

+++ modern-observability-in-practice/background.sh
@@ -5,14 +5,18 @@ curl -s https://datadoghq.dev/katacodalabtools/r?raw=true|bash
 touch /root/status.txt
 sleep 1
 STATUS=$(cat /root/status.txt)
-echo ""> /root/status.txt
 
-mkdir /ecommworkshop
-git clone https://github.com/DataDog/ecommerce-workshop /ecommworkshop
-cd ../ecommworkshop
+if [ "$STATUS" != "complete" ]; then
+  echo ""> /root/status.txt
+  mkdir /ecommworkshop
+  git clone https://github.com/DataDog/ecommerce-workshop /ecommworkshop
+  cd ../ecommworkshop
 
-# locked to specific commit on 2020-10-02
-git checkout 9ce34516d9a65d6f09a6fffd5c4911a409d31e3f
-git reset --hard
+  # locked to specific commit on 2020-10-02
+  git checkout 9ce34516d9a65d6f09a6fffd5c4911a409d31e3f
+  git reset --hard
 
-echo "complete">>/root/status.txt
+  echo "complete">>/root/status.txt
+if
+
+echo "Ready!"
